[Build fixes] Eliminated compilation warnings.
git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@3072 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ orig/trunk/vnc_winsrc/winvnc/vncService.cpp
@@ -505,7 +505,7 @@ vncService::FindWindowByTitle(char *substr)
 	strncpy(l_substr, substr, 255);
 	l_substr[255] = 0;
 	int i;
-	for (i = 0; i < strlen(substr); i++) {
+	for (i = 0; i < (int)strlen(substr); i++) {
 		l_substr[i] = tolower(l_substr[i]);
 	}
